Separate `PydanticAutoDirective` and `PydanticAutoDoc`.
